requirements.txt
streamlit==1.39.0
pillow==10.4.0
opencv-python==4.10.0.84
pyzbar==0.1.9
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0.post0
beautifulsoup4
requests
pyyaml
streamlit-authenticator